% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.points.r
\name{add.points}
\alias{add.points}
\title{add.points}
\usage{
add.points(df, basemap.Info, lat.field = "LATITUDE",
  long.field = "LONGITUDE", plot.field = "EST_COMBINED_WT.SUM",
  plot.field.pretty = NULL, show.legend = T, pnt.style = 19,
  nclasses = 5, cexMax = 3)
}
\arguments{
\item{\code{df}}{the data to be aggregated}

\item{\code{basemap.Info}}{a SpatialPolygons object (identifying the boundaries and projection of an existing plot)}

\item{\code{lat.field}}{the default is \code{'LATITUDE'}. the name of the field holding latitude values (in decimal degrees)}

\item{\code{long.field}}{the default is \code{'LONGITUDE'}.  the name of the field holding longitude values (in decimal degrees)}

\item{\code{plot.field}}{the default is \code{'EST_COMBINED_WT.SUM'}.  The field on which to symbolize the data.}

\item{\code{plot.field.pretty}}{the default is \code{NULL}. A nice name to describe \code{plot.field} (used by the legend)}

\item{\code{show.legend}}{the default is \code{T}. Displays a legend for the added points.}

\item{\code{pnt.style}}{the default is \code{19}. Determines the style of the points (any valid value for \code{pch} is acceptable)}

\item{\code{nclasses}}{the default is \code{5}. Determines how many "bins" to use to display the data.}

\item{\code{cexMax}}{the default is \code{3}. Determines the maximum size of the points to be drawn (any valid value for \code{cex} is 
acceptable)}
}
\value{
a data frame 'allCatchInfo'
}
\description{
T
}
\examples{
 add.points(df, map, lat.field='LAT', lon.field='LON', plot.field='WEIGHT_SMOOTHED', plot.field.pretty='WEIGHT', classes=10, cexMax=1.5)
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\seealso{
Other plotting: \code{\link{get.DFO.areas}}
}

