% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_points.r
\name{add_points}
\alias{add_points}
\title{add_points}
\usage{
add_points(df, basemap = NULL, lat.field = "LATITUDE",
  lon.field = "LONGITUDE", plot.field = NULL, plot.field.pretty = NULL,
  pnt.style = 21, pnt.fill = "#1874CD", pnt.outline = "#333333",
  pnt.cex.min = 1, pnt.cex.max = 2, show.legend = TRUE,
  leg.pos = "topleft", colour.ramp = c("#edf8b1", "#7fcdbb", "#2c7fb8"),
  size.ramp = TRUE, bin.style = "fixed", fixed.breaks.bins = c(1, 4, 32,
  256, 2045, 16336, 130496, 1042432, 8327186), trim.fixed.breaks = TRUE,
  trim.legend = TRUE, nbins = 3, save.plot = FALSE)
}
\arguments{
\item{df}{the dataframe to be plotted}

\item{basemap}{a SpatialPolygons object (identifying the boundaries and projection of an existing plot). If none is provided, a default \code{make_basemap()} object will be used. object}

\item{lat.field}{the default is \code{'LATITUDE'}. The name of the field holding latitude values (in decimal degrees).}

\item{lon.field}{the default is \code{'LONGITUDE'}.  The name of the field holding longitude values (in decimal degrees).}

\item{plot.field}{the default is \code{'EST_COMBINED_WT'}.  The field on which to symbolize the data.  If unspecified, the data can only be symbolized generically.}

\item{plot.field.pretty}{the default is \code{NULL}.  Applies only when \code{bin.style = F}. This is a nice name to describe \code{plot.field} which will be displayed in the legend.}

\item{pnt.style}{the default is \code{21}. Determines the style of the points (any valid value for \code{pch} is acceptable).}

\item{pnt.fill}{the default is \code{'red'}.  Applies only when \code{show.legend = T}.  Determines the colour of the points.}

\item{pnt.outline}{the default is \code{'black'}. This determines the colour outlining the points.}

\item{pnt.cex.min}{the default is \code{1}. When \code{bin.style = F}, this determines the size of the points, and when \code{bin.style = T}, this determines the minimum size of the points to be drawn.}

\item{pnt.cex.max}{the default is \code{2}. When \code{bin.style = F}, this is ignored, and when \code{bin.style = T}, this determines the maximum size of the points to be drawn.}

\item{show.legend}{the default is \code{FALSE}.  Determines whether or not a legend will be displayed.}

\item{leg.pos}{the default is \code{'topleft'}.  Determines where on the plot the legend will be displayed (if \code{show.legend=TRUE}). Valid values are 'topleft','topright','bottomleft','bottomright'.}

\item{colour.ramp}{the default is \code{c("#edf8b1", "#7fcdbb", "#2c7fb8")}, which is color-blind friendly. If this vector exists, data will be symbolized along a gradient of colours as the values in `plot.field` increase. Otherwise, all data will be symbolized as the value selected for `pnt.fill`. You can use hex values, or colour names (e.g. \code{c('blue','pink')}), but there must be at least 2 colour present.}

\item{size.ramp}{the default is \code{TRUE} this indicates whether or not the point size should increase with increasing values of \code{plot.field}}

\item{bin.style}{the default is \code{fixed} chosen style: one of "NULL",  fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", or "jenks".  If \code{NULL}, all points will be displayed  identically, but if a grouping style is selected, the data points will be grouped according to the values in \code{plot.field}.}

\item{fixed.breaks.bins}{the default is \code{c(1,100,1000,100000)} if the \code{bin.style} is set to "fixed", a vector of the desired upper ends of the desired breaks must be supplied (e.g. \code{c(2, 5, 10, 100, 500.10000, 100000)})}

\item{trim.fixed.breaks}{the default is \code{TRUE}  In cases where the \code{fixed.breaks.bins} vector extends beyond the data values, this parameter removes unused categories so that the largest values receive the maximum symbology (i.e. largest pnt.cex.max and extreme colour.ramp colour)}

\item{trim.legend}{the default is \code{TRUE} This ensure that only values that exist in the data are shown in the legend.}

\item{nbins}{the default is \code{3}. Applies only when a \code{bin.style} is selected.  Determines how many "bins" to use to display the data.}

\item{save.plot}{the default is \code{FALSE}.  If FALSE, the plot is displayed, if TRUE, it is saved to the working directory as a bio.plotting.png.}
}
\value{
a SpatialPolygons object  (identifying the boundaries and projection of an existing plot).  It also notifies the user of how many positions lay outside of the map boundaries.
}
\description{
This function adds points to an existing plot.  The advantage over \code{points(...)} is that by including a map object as
input, it will automatically reproject the points, and trim them such that only positions that fall within the map boundaries are shown.
The user is notified of how many positions lay outside of the map boundaries.
Additionally, the input data can either be symbolized generically (i.e. one style), or if a \code{bin.style} is elected, data will be
symbolized according to the values of an identified field.  Both the colour and size of the points can vary with the values in \code{plot.field}.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\seealso{
Other plotting: \code{\link{df_qc_spatial}},
  \code{\link{make_basemap}}
}

