% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.DFO.areas.musquash.r
\name{get.DFO.areas.musquash}
\alias{get.DFO.areas.musquash}
\title{get.DFO.areas.musquash}
\usage{
get.DFO.areas.musquash(DFO.detailed = F)
}
\arguments{
\item{\code{get.detailed}}{default is \code{F}.  Some areas are complex and contains sub-areas, and may not be appropriate for all plots.
\code{get.detailed =T} will plot all of complexity, while \code{get.detailed = F} will only plot the outlines.}
}
\value{
a SpatialPolygon
}
\description{
This function returns SpatialPolygons object of Musquash.  The complexity of the Musquash shape made it easier to separate 
it from the more general \code{get.DFO.areas()} function.
}
\note{
Coordinates for this polygon were given to 14 decimal places, suggesting accuracy to half the thickness of an atom,
I've rounded them all down to 7 decimal places, generously assuming positions were known to 11mm
}
\examples{
 plot(get.DFO.areas.musquash())
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\seealso{
Other plotting: \code{\link{add.points}},
  \code{\link{get.DFO.areas}},
  \code{\link{load.environment}},
  \code{\link{make.basemap}}
}

