% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.DFO.areas.r
\name{get.DFO.areas}
\alias{get.DFO.areas}
\title{get.DFO.areas}
\usage{
get.DFO.areas(requested = c("All"), get.detailed = F)
}
\arguments{
\item{\code{requested}}{default is \code{"All"}. Valid values include the following:
\itemize{
  \item Haddock_Closed_Area
  \item Vazella_Emerald
  \item Vazella_Sambro
  \item Lophelia
  \item Gully
  \item NE_Channel
  \item St_Ann
  \item FundyWhale
  \item RosewayWhale
  \item Haldimand
  \item Shortland
  \item All (plots everything above)
  \item MPA (just plots Gully and Musquash*)
  \item CCA (just plots Vazella_Emerald, Vazella_Sambro, NE_Channel and Gully)
}}

\item{\code{get.detailed}}{default is \code{F}.  Some areas are complex and contains sub-areas, and may not be appropriate for all plots.
\code{get.detailed =T} will plot all of complexity, while \code{get.detailed = F} will only plot the outlines.}
}
\value{
list of SpatialPolygons
}
\description{
This function returns a list of requested SpatialPolygons of areas of interest to DFO.
}
\note{
Musquash is a very complex area, and has a separate function has not been enabled yet.
}
\examples{
 plot(get.DFO.areas('Gully')[[1]])
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\seealso{
Other plotting: \code{\link{add.points}},
  \code{\link{get.DFO.areas.musquash}},
  \code{\link{make.basemap}}
}

