% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_known_areas.r
\name{get_known_areas}
\alias{get_known_areas}
\title{get_known_areas}
\usage{
get_known_areas(known.areas = NULL, known.areas.detailed = NULL)
}
\arguments{
\item{known.areas}{default is \code{NULL}. Valid values include the following:
\itemize{
  \item Bowtie
  \item LFA40
  \item Haddock_Closed_Area
  \item Vazella_Emerald
  \item Vazella_Sambro
  \item Lophelia
  \item Right_Whale_Habitat
  \item Gully
  \item Bottlenose_Habitat
  \item NE_Channel
  \item St_Ann
  \item Musquash
  \item DFO_Regions
  \item DFO_Regions_Lines
  \item US_strata
  \item Mar_strata
  \item NAFO
}
Valid groups include:
\itemize{
  \item mpa (Gully and Musquash)
  \item cca (Vazella_Emerald, Vazella_Sambro, NE_Channel and Gully)
  \item critical_habitat (Bottlenose and Righ Whale Critical Habitats)
  \item Strata (US and Maritimes strata)
}}

\item{known.areas.detailed}{default is \code{F}.  Some areas are complex and contains sub-areas, and may not be appropriate for all plots.
\code{known.areas.detailed  =T} will plot all of complexity, while \code{known.areas.detailed  = F} will only plot the outlines.}
}
\value{
list of SpatialPolygons
}
\description{
This function facilitates the addition of a number of definitive/authoritative/canonical spatial areas.  These may be 
fishing areas, habitat areas, boundaries, etc.  The important thing is that the areas are from a definitive source.  
Multiple areas can be specified, and the \code{known.areas.detailed} flag can be set to either show only the outer boundary, or the 
detailed polygons within it.  Areas can be specified by individually by name, or they can be included in a "group", in which case, 
calling the group will return all group members.  An individual layer can be in many groups.  Area identified are case-insensitive.
}
\note{
As it is now written, this function references paths that will not be visible to most people.  The intent is that these canonical
data layers will be moved to a location accessible by all staff.
}
\examples{
 this.is.the.gully = get_known_areas('Gully')[[1]]
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\seealso{
Other plotting: \code{\link{add_points}},
  \code{\link{load_plotting_environment}},
  \code{\link{make_basemap}}
}

