% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.r
\docType{data}
\name{isobaths}
\alias{isobaths}
\title{isobaths}
\description{
Includes isobaths for the following depths (m):
 "10"   "20"   "50"   "75"   "100"  "200"  "250"  "300"  "400"  "500"  "600"  "700"  "750"  "800"
 "900"  "1000" "1200" "1250" "1400" "1500" "1750" "2000" "2500" "3000" "4000" "5000"
}
\author{
Jae Choi, \email{Jae.Choi@dfo-mpo.gc.ca}
}
\references{
\url{data_blah.com}
}
\keyword{data}

