% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.environment.r
\name{load.environment}
\alias{load.environment}
\title{load.environment}
\usage{
load.environment(libs = NULL, p.plotting = NULL)
}
\value{
a list of parameters - \code{p.plotting}'
}
\description{
This function sets a bunch of parameters and outlines the package requirements for the bio.plotting package
This way, a variety of functions can use them without having to pass them through function calls
}
\examples{
 p.plotting = bio.plotting::load.environment() 
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\seealso{
Other plotting: \code{\link{add.points}},
  \code{\link{get.DFO.areas.musquash}},
  \code{\link{get.DFO.areas}}, \code{\link{make.basemap}}
}

