% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_plotting_environment.r
\name{load_plotting_environment}
\alias{load_plotting_environment}
\title{load_plotting_environment}
\usage{
load_plotting_environment()
}
\value{
a list of parameters - \code{p.plotting}
}
\description{
This function sets a bunch of parameters and outlines the package requirements for the bio.plotting package
This way, a variety of functions can use them without having to pass them through function calls
}
\examples{
 p.plotting = load_plotting_environment()
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\seealso{
Other plotting: \code{\link{add_points}},
  \code{\link{make_basemap}}
}

