% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap.r
\name{qmap}
\alias{qmap}
\title{qmap}
\usage{
qmap(db, save.plot = F)
}
\arguments{
\item{db}{This identifies the dataset you want to plot.  Valid values are 'obs'
(observer) and 'rv' (groundfish/research survey).}

\item{save.plot}{the default is \code{FALSE}.  If FALSE, the plot is displayed, if TRUE, it is saved to the working directory as a bio.plotting.png.}
}
\value{
list of the filters that were applied to generate the remaining dataset
}
\description{
qmap ('quickmap') is a simple script for quickly generating plots from bio.datawrangling datasets.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}

