% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_points.r
\name{add_points}
\alias{add_points}
\title{add_points}
\usage{
add_points(df, basemap = NULL, lat.field = "LATITUDE",
  lon.field = "LONGITUDE", plot.field = NULL, plot.field.pretty = NULL,
  pnt.style = 21, pnt.col = "black", pnt.bg = "red", pnt.cex.min = 1,
  pnt.cex.max = 2, show.legend = TRUE, leg.pos = "topleft",
  use.buckets = TRUE, use.colours = TRUE, colour.ramp = c("#edf8b1",
  "#7fcdbb", "#2c7fb8"), bucket.style = "quantile",
  bucket.fixed.breaks = NULL, nclasses = 3, save.plot = FALSE)
}
\arguments{
\item{df}{the dataframe to be plotted}

\item{basemap}{a SpatialPolygons object (identifying the boundaries and projection of an existing plot). If none is provided, a default \code{make_basemap()} object will be used. object}

\item{lat.field}{the default is \code{'LATITUDE'}. The name of the field holding latitude values (in decimal degrees).}

\item{lon.field}{the default is \code{'LONGITUDE'}.  The name of the field holding longitude values (in decimal degrees).}

\item{plot.field}{the default is \code{'EST_COMBINED_WT'}.  The field on which to symbolize the data.  If unspecified, the data can only be symbolized generically.}

\item{plot.field.pretty}{the default is \code{NULL}.  Applies only when \code{use.buckets = F}. This is a nice name to describe \code{plot.field} which will be displayed in the legend.}

\item{pnt.style}{the default is \code{21}. Determines the style of the points (any valid value for \code{pch} is acceptable).}

\item{pnt.col}{the default is \code{'black'}. This determines the colour outlining the points.}

\item{pnt.bg}{the default is \code{'red'}.  Applies only when \code{show.legend = T}.  Determines the colour of the points.}

\item{pnt.cex.min}{the default is \code{1}. When \code{use.buckets = F}, this determines the size of the points, and when \code{use.buckets = T}, this determines the minimum size of the points to be drawn.}

\item{pnt.cex.max}{the default is \code{2}. When \code{use.buckets = F}, this is ignored, and when \code{use.buckets = T}, this determines the maximum size of the points to be drawn.}

\item{show.legend}{the default is \code{FALSE}.  Determines whether or not a legend will be displayed.}

\item{leg.pos}{the default is \code{'topleft'}.  Determines where on the plot the legend will be displayed (if \code{show.legend=TRUE}). Valid values are 'topleft','topright','bottomleft','bottomright'.}

\item{use.buckets}{the default is \code{TRUE}. If \code{use.buckets = F}, all points are identical, but if \code{use.buckets = T}, data points are scaled, and the colour intensity varies according to the value of \code{plot.field}.}

\item{use.colours}{the default is \code{TRUE}. If \code{use.colour = FALSE}, all points will be coloured using the value of \code{pnt.bg}.  If set to \code{TRUE}, the the intensity of the colour will also scale with the size of the markers, according to the value of \code{plot.field}.}

\item{colour.ramp}{the default is \code{c("#edf8b1", "#7fcdbb", "#2c7fb8")}, which is color-blind friendly. You can use hex values, or colour names (e.g. \code{c('blue','pink')}), but there must be at least 2 colour present.}

\item{bucket.style}{the default is \code{quantile} chosen style: one of "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", or "jenks"}

\item{bucket.fixed.breaks}{the default is \code{NULL} if the \code{bucket.style} is set to "fixed", a vector of the desired upper ends of the desired breaks must be supplied (e.g. \code{c(2, 5, 10, 100, 500.10000, 100000)})}

\item{nclasses}{the default is \code{3}. Applies only when \code{use.buckets = T}.  Determines how many "bins" to use to display the data.}

\item{save.plot}{the default is \code{FALSE}.  If FALSE, the plot is displayed, if TRUE, it is saved to the working directory as a bio.plotting.png.}
}
\value{
NULL, but notifies the user of how many positions lay outside of the map boundaries.
}
\description{
This function adds points to an existing plot.  The advantage over \code{points(...)} is that by including a map object as
input, it will automatically reproject the points, and trim them such that only positions that fall within the map boundaries are shown.
The user is notified of how many positions lay outside of the map boundaries.
Additionally, the input data can either be symbolized generically (i.e. one style), or it can classify the points according to the values
stored in a field identified by the user using the flag \code{use.buckets}.  If \code{use.buckets=T}, data is symbolized by a selected
field, and the colour and size of the points varies with the values in \code{plot.field}.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\seealso{
Other plotting: \code{\link{make_basemap}}
}

